% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleAge.R
\name{adjustAgeDistribution}
\alias{adjustAgeDistribution}
\title{Adjust cumulative age distribution to match target mean.}
\usage{
adjustAgeDistribution(AgeCDF_Ag, TargetMean = NULL)
}
\arguments{
\item{AgeCDF_Ag}{A named numeric vector where the names are vehicle ages and
the values are the proportion of vehicles that age or younger. The names must
be an ordered sequence from 0 to 30.}

\item{TargetMean}{A number that is the target mean value.}
}
\value{
A numeric value that is the mean vehicle age.
}
\description{
\code{adjustAgeDistribution} Adjusts a cumulative age distribution to match a
target mean age.
}
\details{
This function adjusts a cumulative age distribution to match a target mean
age. The function returns the adjusted cumulative age distribution and the
corresponding age distribution. If no target mean value is specified, the
function returns the input cumulative age distribution and the corresponding
age distribution for that input.
}
