% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignDrivers.R
\name{AssignDrivers}
\alias{AssignDrivers}
\title{Main module function to assign drivers by age group to each household.}
\usage{
AssignDrivers(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignDrivers} assigns number of drivers by age group to each household.
}
\details{
This function assigns the number of drivers in each age group to each
household. It also computes the total number of drivers in the household.
}
