% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleFeatures.R
\name{assignFuelEconomy}
\alias{assignFuelEconomy}
\title{Assignes mileage to vehicles in a household}
\usage{
assignFuelEconomy(Hh_df, VehMpgYr = NULL, CurrentYear)
}
\arguments{
\item{Hh_df}{A household data frame consisting of household characteristics.}

\item{VehMpgYr}{A data frame of mileage of vehicles by type and year.}

\item{CurrentYear}{A integer indicating the current year.}
}
\value{
A numeric vector that indicates the mileage of vehicles.
}
\description{
\code{assignFuelEconomy} Assignes mileage to vehicles in a household.
}
\details{
This function assigns mileage to vehicles in a household based on type
age of the vehicles.
}
