% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleAge.R
\name{AssignVehicleAge}
\alias{AssignVehicleAge}
\title{Create vehicle table and populate with vehicle type and age records.}
\usage{
AssignVehicleAge(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignVehicleAge} create the vehicle table and populate with vehicle
age and type records.
}
\details{
This function creates the 'Vehicle' table in the datastore and populates it
with records of vehicle types and ages along with household IDs.
}
