% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleAge.R
\docType{data}
\name{AssignVehicleAgeSpecifications}
\alias{AssignVehicleAgeSpecifications}
\title{Specifications list for AssignVehicleAge module}
\format{A list containing 5 components:
\describe{
 \item{NewSetTable}{table to be created}
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{model inputs to be saved to the datastore}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
AssignVehicleAge.R script.
}
\usage{
AssignVehicleAgeSpecifications
}
\description{
A list containing specifications for the AssignVehicleAge module.
}
\keyword{datasets}
