% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleFeatures.R
\name{AssignVehicleFeatures}
\alias{AssignVehicleFeatures}
\title{Create vehicle table and populate with vehicle type, age, and mileage records.}
\usage{
AssignVehicleFeatures(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignVehicleFeatures} create vehicle table and populate with
vehicle type, age, and mileage records.
}
\details{
This function creates the 'Vehicle' table in the datastore and populates it
with records of vehicle types, ages, mileage, and mileage proportions
along with household IDs.
}
