% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleFeaturesFuture.R
\name{AssignVehicleFeaturesFuture}
\alias{AssignVehicleFeaturesFuture}
\title{Create vehicle table and populate with vehicle type, age, and mileage records.}
\usage{
AssignVehicleFeaturesFuture(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignVehicleFeaturesFuture} populate vehicle table with
vehicle type, age, and mileage records using future data.
}
\details{
This function populates vehicle table with records of
vehicle types, ages, mileage, and mileage proportions
along with household IDs using future data.
}
