% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleFeaturesFuture.R
\docType{data}
\name{AssignVehicleFeaturesFutureSpecifications}
\alias{AssignVehicleFeaturesFutureSpecifications}
\title{Specifications list for AssignVehicleFeaturesFuture module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
AssignVehicleFeaturesFuture.R script.
}
\usage{
AssignVehicleFeaturesFutureSpecifications
}
\description{
A list containing specifications for the AssignVehicleFeaturesFuture module.
}
\keyword{datasets}
