% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleFeatures.R
\docType{data}
\name{AssignVehicleFeaturesSpecifications}
\alias{AssignVehicleFeaturesSpecifications}
\title{Specifications list for AssignVehicleFeatures module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
AssignVehicleFeatures.R script.
}
\usage{
AssignVehicleFeaturesSpecifications
}
\description{
A list containing specifications for the AssignVehicleFeatures module.
}
\keyword{datasets}
