% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleOwnership.R
\name{AssignVehicleOwnership}
\alias{AssignVehicleOwnership}
\title{Calculate the number of vehicles owned by the household.}
\usage{
AssignVehicleOwnership(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignVehicleOwnership} calculate the number of vehicles owned by each
household.
}
\details{
This function calculates the number of vehicles owned by each household
given the characteristic of the household and the area where it resides.
}
