% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleOwnership.R
\docType{data}
\name{AssignVehicleOwnershipSpecifications}
\alias{AssignVehicleOwnershipSpecifications}
\title{Specifications list for AssignVehicleOwnership module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
AssignVehicleOwnership.R script.
}
\usage{
AssignVehicleOwnershipSpecifications
}
\description{
A list containing specifications for the AssignVehicleOwnership module.
}
\keyword{datasets}
