% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehiclePowertrain.R
\docType{data}
\name{AssignVehiclePowertrainSpecifications}
\alias{AssignVehiclePowertrainSpecifications}
\title{Specifications list for AssignVehiclePowertrain module}
\format{A list containing 4 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{model inputs to be saved to the datastore}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
AssignVehiclePowertrain.R script.
}
\usage{
AssignVehiclePowertrainSpecifications
}
\description{
A list containing specifications for the AssignVehiclePowertrain module.
}
\keyword{datasets}
