% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleType.R
\name{AssignVehicleType}
\alias{AssignVehicleType}
\title{Assign number of autos and light trucks for each household.}
\usage{
AssignVehicleType(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignVehicleType} assigns the numbers of autos and light trucks in
each household.
}
\details{
This function assigns the numbers of autos and light trucks in each
household.
}
