% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleType.R
\docType{data}
\name{AssignVehicleTypeSpecifications}
\alias{AssignVehicleTypeSpecifications}
\title{Specifications list for AssignVehicleType module}
\format{A list containing 5 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{model inputs to be saved to the datastore}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
AssignVehicleType.R script.
}
\usage{
AssignVehicleTypeSpecifications
}
\description{
A list containing specifications for the AssignVehicleType module.
}
\keyword{datasets}
