% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleOwnership.R
\docType{data}
\name{AutoOwnModels_ls}
\alias{AutoOwnModels_ls}
\title{Auto ownership model}
\format{A list having the following components:
\describe{
  \item{Metro}{a list containing two models for metropolitan areas: a Zero
  component that is a binomial logit model for determining which households
  own no vehicles and a Count component that is an ordered logit model for
  determining how many vehicles a household who has vehicles owns}
  \item{NonMetro}{a list containing two models for non-metropolitan areas: a
  Zero component that is a binomial logit model for determining which households
  own no vehicles and a Count component that is an ordered logit model for
  determining how many vehicles a household who has vehicles owns}
}}
\source{
AssignVehicleOwnership.R script.
}
\usage{
AutoOwnModels_ls
}
\description{
A list containing the auto ownership model equation and other information
needed to implement the auto ownership model.
}
\keyword{datasets}
