% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleAge.R
\name{calcAgeDistributionByInc}
\alias{calcAgeDistributionByInc}
\title{Calculate vehicle age distributions by income group.}
\usage{
calcAgeDistributionByInc(Seed_AgIg, Margin_Ag, Margin_Ig, MaxIter = 100,
  Closure = 1e-04)
}
\arguments{
\item{Seed_AgIg}{A numeric matrix of the joint probabilities of vehicles
by age and income group.}

\item{Margin_Ag}{A numeric vector of vehicle age probabilities.}

\item{Margin_Ig}{A numeric vector of vehicle household income probabilities.}

\item{MaxIter}{A numeric value specifying the maximum number of iterations
the iterative proportional fitting process will undertake.}

\item{Closure}{A numeric value specifying the maximum allowed difference
between any margin value and corresponding sum of values of the joint
probability matrix.}
}
\value{
A numeric value that is the mean vehicle age.
}
\description{
\code{calcAgeDistributionByInc} Calculates vehicle age distributions by
household income group.
}
\details{
This function calculates vehicle age distributions by household income group.
It takes marginal distributions of vehicles by age and vehicles by household
income group along with a seed matrix of the joint probability distribution
of vehicles by age and income group, and then uses iterative proportional
fitting to adjust the joint probabilities to match the margins. The
probabilities by income group are calculated from the fitted joint
probability matrix. The seed matrix is the joint age and income distribution
for autos or light trucks in the VehicleAgeModel_ls (AgeIncJointProp_AgIg).
The age margin is the proportional distribution of vehicles by age calculated
by adjusting the cumulative age distribution for autos or light trucks in the
VehicleAgeModel_ls (AgeCDF_AgTy) to match a target mean age. The income
margin is the proportional distribution of vehicles by household income group
($0-20K, $20K-40K, $40K-60K, $60K-80K, $80K-100K, $100K or more) calculated
from the modeled household values.
}
