% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateVehicleOwnCost.R
\docType{data}
\name{CalculateVehicleOwnCostSpecifications}
\alias{CalculateVehicleOwnCostSpecifications}
\title{Specifications list for CalculateVehicleOwnCost module}
\format{A list containing 5 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{model inputs to be saved to the datastore}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
 \item{Call}{alias and name of module to be called}
}}
\source{
CalculateVehicleOwnCost.R script.
}
\usage{
CalculateVehicleOwnCostSpecifications
}
\description{
A list containing specifications for the CalculateVehicleOwnCost module.
}
\keyword{datasets}
