% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateVehicleOwnCost.R
\name{calcVehDepr}
\alias{calcVehDepr}
\title{Calculate vehicle depreciation}
\usage{
calcVehDepr(Type_, Age_, Vmt_)
}
\arguments{
\item{Type_}{A character vector of vehicle types (Auto, LtTrk)}

\item{Age_}{A numeric vector of vehicle ages}

\item{Vmt_}{A numeric vector of the annual vehicle miles traveled for the
vehicles}
}
\value{
A numeric vector of annual depreciation cost in 2017 dollars
}
\description{
\code{calcVehDepr} calculates vehicle depreciation given vehicle type, age,
and annual mileage
}
\details{
This function calculates the annual depreciation cost (in 2017 dollars) of
vehicles as a function of the vehicle type (Auto, LtTrk), age, and annual
mileage. A base depreciation value is calculated using the depreciation cost
matrix (VehOwnCost_ls$Depr_AgVt) calculated from AAA data in the module
script. The base depreciation is a function of vehicle type and age. The
base depreciation is adjusted based on the vehicle's annual mileage using the
depreciation adjustment models (VehOwnCost_ls$DeprAdjModel_ls). The models,
one for each vehicle type (Auto, LtTrk) are quadratic polynomials with
minimum values at 10,000 miles so the minimum vehicle VMT is constrained to
10,000 miles for use in the model.
}
