% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateVehicleOwnCost.R
\name{calcVehFin}
\alias{calcVehFin}
\title{Calculate vehicle finance cost}
\usage{
calcVehFin(Type_, Age_)
}
\arguments{
\item{Type_}{A character vector of vehicle types (Auto, LtTrk)}

\item{Age_}{A numeric vector of vehicle ages
vehicles}
}
\value{
A numeric vector of annual finance cost in 2017 dollars
}
\description{
\code{calcVehFin} calculates vehicle finance cost given vehicle type and age
}
\details{
This function calculates the annual financing cost (in 2017 dollars) of
vehicles as a function of the vehicle type (Auto, LtTrk) and age using the ,
and annual finance cost matrix (VehOwnCost_ls$FinCost_AgVt) calculated from
AAA data in the module script.
}
