% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleFeatures.R
\name{calcVehicleAges}
\alias{calcVehicleAges}
\title{Calculate vehicle type and age for each household.}
\usage{
calcVehicleAges(Hh_df, VProp = NULL, AdjRatio = c(Auto = 1, LtTruck = 1))
}
\arguments{
\item{Hh_df}{A household data frame consisting of household characteristics.}

\item{VProp}{A list consisting of a cumulative distribution of vehicle age by
vehicle type and a joint distribution of vehicle age, type and income group of
the household.}

\item{AdjRatio}{A number that is the target ratio value.}
}
\value{
A list containing the vehicle types and ages for each household.
}
\description{
\code{calcVehicleAges} Calculates vehicle type and age for each household.
}
\details{
This function calculates the vehicle type and age for households. The function
uses characteristics of houshold and target marginal proportions to calculate
vehicle type and age.
}
