% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleFeatures.R
\name{calcVehPropByIncome}
\alias{calcVehPropByIncome}
\title{Calculate vehicle type distributions by income group.}
\usage{
calcVehPropByIncome(Hh_df)
}
\arguments{
\item{Hh_df}{A household data frame consisting of household characteristics.}
}
\value{
A data frame containing the distribution of vehicle types by income
groups.
}
\description{
\code{calcVehPropByIncome} Calculates vehicle type distributions by
household income group.
}
\details{
This function calculates vehicle type distributions by household
income group. It takes the the number of vehicles, vehicle types, and
income groups of each household and calculates the marginal distribution
of the vehicle types.
}
