% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVehicleTable.R
\name{CreateVehicleTable}
\alias{CreateVehicleTable}
\title{Create vehicle table and populate with HhId and Azone datasets.}
\usage{
CreateVehicleTable(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CreateVehicleTable} create the vehicle table and populate with HhId
and Azone datasets.
}
\details{
This function creates the 'Vehicle' table in the datastore and populates it
with HhId and Azone datasets.
}
