% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateVehicleTable.R
\docType{data}
\name{CreateVehicleTableSpecifications}
\alias{CreateVehicleTableSpecifications}
\title{Specifications list for CreateVehicleTable module}
\format{A list containing 5 components:
\describe{
 \item{NewSetTable}{table to be created}
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{model inputs to be saved to the datastore}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
CreateVehicleTable.R script.
}
\usage{
CreateVehicleTableSpecifications
}
\description{
A list containing specifications for the CreateVehicleTable module.
}
\keyword{datasets}
