% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignDrivers.R
\docType{data}
\name{DriverModel_ls}
\alias{DriverModel_ls}
\title{Driver choice model}
\format{A list having having Metro and NonMetro components, with each having the following components:
\describe{
  \item{Type}{a string identifying the type of model}
  \item{Formula}{a string representation of the model formula}
  \item{PrepFun}{a function that prepares inputs to be applied in the model}
  \item{Summary}{the summary of the binomial logit model estimation results}
  \item{Anova}{results of analysis of variance of the model}
  \item{PropCorrectlyPredicted}{proportion of cases of validation dataset correctly predicted by model}
  \item{SearchRange}{a two-element vector specifying the range of search values}
}}
\source{
AssignDrivers.R script.
}
\usage{
DriverModel_ls
}
\description{
A list containing the driver choice models for metropolitan and
non-metropolitan areas. Includes model equations and other information
needed to implement the driver choice model.
}
\keyword{datasets}
