% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignDrivers.R
\name{estimateDriverModel}
\alias{estimateDriverModel}
\title{Estimate housing choice model}
\usage{
estimateDriverModel(Data_df, StartTerms_, ValidationProp)
}
\arguments{
\item{Data_df}{A data frame containing estimation data.}

\item{StartTerms_}{A character vector of the terms of the model to be
tested in the model.}

\item{ValidationProp}{A numeric value between 0 and 0.5 which specified
the proportion of Data_df that will be the validation dataset. The remainder
will be the training dataset.}
}
\value{
A list which has the following components:
Type: a string identifying the type of model ("binomial"),
Formula: a string representation of the model equation,
PrepFun: a function that prepares inputs to be applied in the binomial model,
OutFun: a function that transforms the result of applying the binomial model.
Summary: the summary of the binomial model estimation results.
Anova: the results of analysis of variance of the model.
PropCorrectlyPredicted: the proportion of validation set observed values that
are correctly predicted by the estimated. The model is estimated using 80% of
the data and validated using the remaining 20%. The proportion of correct
predictions for the validation data set is calculated.
}
\description{
\code{estimateDriverModel} estimates a binomial logit model for choosing
between driver and non-driver
}
\details{
This function estimates a binomial logit model for predicting whether a
person is a driver or a non-driver
}
