% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleType.R
\name{estimateVehicleTypeModel}
\alias{estimateVehicleTypeModel}
\title{Estimate vehicle type choice model}
\usage{
estimateVehicleTypeModel(EstData_df, Counts_mx, StartTerms_)
}
\arguments{
\item{EstData_df}{A data frame containing estimation data.}

\item{Counts_mx}{A numeric matrix of counts of household light trucks and
automobiles for each household.}

\item{StartTerms_}{A character vector of the terms of the model to be
tested in the model. The function estimates the model using these terms
and then drops all terms whose p value is greater than 0.05.}
}
\value{
A list which has the following components:
Type: a string identifying the type of model ("binomial"),
Formula: a string representation of the model equation,
PrepFun: a function that prepares inputs to be applied in the binomial model,
OutFun: a function that transforms the result of applying the binomial model.
Summary: the summary of the binomial model estimation results.
}
\description{
\code{estimateVehicleTypeModel} estimates a binomial logit model for choosing
between light trucks and automobiles.
}
\details{
This function estimates a binomial logit model for predicting vehicle type
choice (automobile or light truck) as a function of the characteristics of
the household, the number of vehicles it owns, the place where the household
resides, and targets for light-truck ownership.
}
