% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleAge.R
\name{findMeanAge}
\alias{findMeanAge}
\title{Calculate mean vehicle age from cumulative age distribution.}
\usage{
findMeanAge(AgeCDF_Ag)
}
\arguments{
\item{AgeCDF_Ag}{A named numeric vector where the names are vehicle ages and
the values are the proportion of vehicles that age or younger. The names must
be an ordered sequence from 0 to 30.}
}
\value{
A numeric value that is the mean vehicle age.
}
\description{
\code{findMeanAge} calculates mean age from a cumulative age distribution.
}
\details{
This function calculates a mean age from a cumulative age distribution vector
where the values of the vector are the cumulative proportions and the names
of the vector are the vehicle ages from 0 to 30 years.
}
