% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateVehicleOwnCost.R
\name{idPaydHh}
\alias{idPaydHh}
\title{Assign pay-as-you-drive insurance propensity weights to households}
\usage{
idPaydHh(Household, PaydHhProp)
}
\arguments{
\item{Household}{A list containing the Household component of the list (L)
including Household table data listed in the Get specifications limited to
data for households in a specified Azone}

\item{PaydHhProp}{A number identifying the proportion of households having
pay-as-you-drive insurance
for the module.}
}
\value{
A numeric vector of weights assigned to each household
}
\description{
\code{calcPaydWeights} Calculates household weight that reflect the relative
propensity of a household to purchase pay-as-you-drive insurance based on the
household characteristics
}
\details{
Household PAYD propensity weights are assigned based on the presence of
teenager drivers, whether the average annual vehicle mileage is low,
the proportion of older drivers in the household,  whether household income
is relatively low, the proportion of household vehicles that are autos, and
whether the household lives in a metropolitan area. All household vehicles
must be a 1996 or later model year.
}
