% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateVehicleOwnCost.R
\docType{data}
\name{PaydWts_}
\alias{PaydWts_}
\title{Household attributes weights for PAYD insurance}
\format{A named vector of weights used for determining household weight for selecting PAYD insurance
\describe{
  \item{HasTeenDrv}{weight for households having one or more teenage drivers},
  \item{LowerMileage}{weight for households driving lower mileage (< 15,000 per vehicle)}
  \item{OlderDrvProp}{weight for proportion of drivers in the household who are 30 or older}
  \item{LowerIncome}{weight for lower income households (< 45,000 year 2005 dollars)}
  \item{AutoProp}{weight for automobile proportion of vehicles owned by household}
  \item{InMetroArea}{weight for household being located in a metropolitan (urbanized) area}
}}
\source{
CalculateVehicleOwnCost.R script.
}
\usage{
PaydWts_
}
\description{
Identifies household attributes associated with higher probability of PAYD
insurance and the relative weights of those attributes.
}
\keyword{datasets}
