% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleFeatures.R
\name{predictLtTruckOwn}
\alias{predictLtTruckOwn}
\title{Predict vehicle type (automobile or light truck) for household vehicles.}
\usage{
predictLtTruckOwn(Hh_df, ModelType = LtTruckModels_ls, TruckProp = NA)
}
\arguments{
\item{Hh_df}{A household data frame consisting of household characteristics.}

\item{ModelType}{A list of light truck ownership model.}

\item{TruckProp}{A numeric indicating the target proportion for light truck
ownership.}
}
\value{
A list containing vehicle types for each household.
}
\description{
\code{predictVehicleOwnership} Predict vehicle type (automobile or light truck)
for household vehicles.
}
\details{
This function predict vehicle type (automobile or light truck)
for household vehicles based on characterisitics of the household, the place where
the household resides, the number of vehicles it owns, and areawide targets for
light truck ownership.
}
