% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleFeatures.R
\name{predictVehicleOwnership}
\alias{predictVehicleOwnership}
\title{Predict vehicle ownership to match the target proportion for a specific
region type.}
\usage{
predictVehicleOwnership(Hh_df, ModelType = VehOwnModels_ls,
  VehProp = NA, Type = "Metro")
}
\arguments{
\item{Hh_df}{A household data frame consisting of household attributes.}

\item{ModelType}{A list of vehicle ownership models.}

\item{VehProp}{A list of data frame consisting of distribution of number
of vehicles by driving age population for each region type.}

\item{Type}{A string indicating the region type ("Metro": Default, or "NonMetro")}
}
\value{
A list containing number of vehicles and ownership ratio for each household
}
\description{
\code{predictVehicleOwnership} Predict vehicle ownership to match the
target proportion for a specific region type.
}
\details{
This function predicts the number of vehicles and the ratio of number of vehicles
to the driving age population (ownership ratio).
}
