% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleAge.R
\docType{data}
\name{VehicleAgeModel_ls}
\alias{VehicleAgeModel_ls}
\title{Vehicle age model}
\format{A list having the following components:
\describe{
  \item{Auto$AgeCDF_Ag}{a vector of cumulative probability of autos by age}
  \item{Auto$AgeIncJointProp_AgIg}{a matrix of the joint probability of autos by age and household income}
  \item{LtTrk$AgeCDF_Ag}{a vector of cumulative probability of light trucks by age}
  \item{LtTrk$AgeIncJointProp_AgIg}{a matrix of the joint probability of light trucks by age and household income}
}}
\source{
AssignVehicleAge.R script.
}
\usage{
VehicleAgeModel_ls
}
\description{
A list containing the vehicle age model probability tables
needed to implement the vehicle age model.
}
\keyword{datasets}
