% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleType.R
\docType{data}
\name{VehicleTypeModel_ls}
\alias{VehicleTypeModel_ls}
\title{Vehicle type choice model}
\format{A list having the following components:
\describe{
  \item{Type}{a string identifying the type of model ("binomial")}
  \item{Formula}{makeModelFormulaString(VehicleTypeModel)}
  \item{PrepFun}{a function that prepares inputs to be applied in the model}
  \item{Summary}{the summary of the binomial logit model estimation results}
  \item{SearchRange}{a two-element vector specifying the range of search values}
}}
\source{
AssignVehicleType.R script.
}
\usage{
VehicleTypeModel_ls
}
\description{
A list containing the vehicle type choice model equation and other information
needed to implement the vehicle type choice model.
}
\keyword{datasets}
