% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleFeatures.R
\docType{data}
\name{VehOwnModels_ls}
\alias{VehOwnModels_ls}
\title{Vehicle ownership model}
\format{A list having the following components:
\describe{
  \item{Metro}{a list containing four models for metropolitan areas: a Zero
  component model and three separate models for non-zero component}
  \item{NonMetro}{a list containing four models for non-metropolitan areas: a
  Zero component model and three separate models for non-zero component}
}}
\source{
AssignVehicleFeatures.R script.
}
\usage{
VehOwnModels_ls
}
\description{
A list containing the vehicle ownership model equation and other information
needed to implement the vehicle ownership model.
}
\keyword{datasets}
