% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignVehicleFeatures.R
\docType{data}
\name{LtTruckModels_ls}
\alias{LtTruckModels_ls}
\title{Light truck ownership model}
\format{A list having the following components:
\describe{
  \item{OwnModel}{The light truck ownership model}
}}
\source{
AssignVehicleFeatures.R script.
}
\usage{
LtTruckModels_ls
}
\description{
A list containing the light truck ownership model equation.
}
\keyword{datasets}
